player clean arch
