+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".AudioPlayerActivity"
+            android:name=".ui.AudioPlayerActivity"
             android:exported="false" />
         <activity
             android:name=".SearchActivity"
