+++ app/src/main/java/com/sakal/playlistmaker/ApiConstants.kt
@@ -1,4 +1,4 @@
-package com.sakal.playlistmaker.model
+package com.sakal.playlistmaker
 
 object ApiConstants {
     const val BASE_URL = "http://itunes.apple.com/"
