--- app/src/main/java/com/sakal/playlistmaker/AudioPlayerActivity.kt
@@ -1,203 +1,0 @@
-package com.sakal.playlistmaker
-
-import android.media.MediaPlayer
-import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-import android.view.View
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.google.gson.Gson
-import com.sakal.playlistmaker.model.Track
-import java.text.SimpleDateFormat
-import java.util.*
-import androidx.appcompat.widget.Toolbar
-
-class AudioPlayerActivity : AppCompatActivity() {
-
-    private lateinit var toolbar: Toolbar
-    private lateinit var trackName: TextView
-    private lateinit var trackTime: TextView
-    private lateinit var artistName: TextView
-    private lateinit var albumIcon: ImageView
-    private lateinit var collectionName: TextView
-    private lateinit var collectionNameTitle: TextView
-    private lateinit var releaseDate: TextView
-    private lateinit var primaryGenreName: TextView
-    private lateinit var country: TextView
-    private lateinit var progress: TextView
-    private lateinit var buttonPlay: FloatingActionButton
-    private val gson = Gson()
-    private val play = R.drawable.play_arrow
-    private val pause = R.drawable.pause
-    private var mediaPlayer = MediaPlayer()
-    private var playerState = STATE_DEFAULT
-    private var mainHandler: Handler? = null
-
-    private val runThread = object : Runnable {
-        override fun run() {
-            progress.text =
-                SimpleDateFormat("mm:ss", Locale.getDefault()).format(mediaPlayer.currentPosition)
-
-            mainHandler?.postDelayed(
-                this,
-                Constants.RELOAD_PROGRESS
-            )
-        }
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_audioplayer)
-
-        initToolbar()
-
-        initTrackInfo()
-
-    }
-
-    private fun initToolbar() {
-        toolbar = findViewById(R.id.player_toolbar)
-        toolbar.setNavigationOnClickListener {
-            finish()
-        }
-    }
-
-    private fun initTrackInfo() {
-        val track = gson.fromJson(
-            intent.getStringExtra(Constants.TRACK),
-            Track::class.java
-        )
-        mainHandler = Handler(Looper.getMainLooper())
-
-        trackName = findViewById(R.id.trackName)
-        artistName = findViewById(R.id.artistName)
-        trackTime = findViewById(R.id.trackTime)
-        albumIcon = findViewById(R.id.track_icon)
-        collectionName = findViewById(R.id.album_name)
-        releaseDate = findViewById(R.id.release_date_data)
-        primaryGenreName = findViewById(R.id.primary_genre_name)
-        country = findViewById(R.id.country_data)
-        buttonPlay = findViewById(R.id.play_track)
-        progress = findViewById(R.id.progress)
-
-
-        Glide
-            .with(albumIcon)
-            .load(track.artworkUrl100.replaceAfterLast('/', "512x512bb.jpg"))
-            .placeholder(R.drawable.placeholder_512)
-            .centerCrop()
-            .transform(
-                RoundedCorners(resources.getDimensionPixelSize(R.dimen.corner_radius_8))
-            )
-            .into(albumIcon)
-
-        trackName.text = track.trackName
-        artistName.text = track.artistName
-        primaryGenreName.text = track.primaryGenreName
-        country.text = track.country
-
-        trackTime.text =
-            SimpleDateFormat("mm:ss", Locale.getDefault()).format(track.trackTimeMillis.toInt())
-
-        val date = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault()).parse(track.releaseDate)
-        if (date != null) {
-            val formatDatesString = SimpleDateFormat("yyyy", Locale.getDefault()).format(date)
-            releaseDate.text = formatDatesString
-        }
-
-        if (track.collectionName.isNotEmpty()) {
-            collectionName.text = track.collectionName
-        } else {
-            collectionName.visibility = View.GONE
-            collectionNameTitle.visibility = View.GONE
-        }
-
-        preparePlayer(track)
-
-        buttonPlay.setOnClickListener {
-            playbackControl()
-            progressTimeControl()
-        }
-
-        mediaPlayer.setOnCompletionListener {
-            playerState = STATE_PREPARED
-            mainHandler?.removeCallbacks(runThread)
-            buttonPlay.setImageResource(play)
-            progress.text = Constants.CURRENT_TIME_ZERO
-        }
-    }
-
-    private fun preparePlayer(track: Track) {
-        mediaPlayer.setDataSource(track.previewUrl)
-        mediaPlayer.prepareAsync()
-        mediaPlayer.setOnPreparedListener {
-            buttonPlay.setImageResource(play)
-            playerState = STATE_PREPARED
-        }
-        mediaPlayer.setOnCompletionListener {
-            buttonPlay.setImageResource(play)
-            playerState = STATE_PREPARED
-        }
-    }
-
-    private fun progressTimeControl() {
-        when (playerState) {
-            STATE_PLAYING -> {
-                mainHandler?.postDelayed(
-                    runThread,
-                    Constants.RELOAD_PROGRESS
-                )
-            }
-            STATE_PAUSED -> {
-                mainHandler?.removeCallbacks(runThread)
-            }
-        }
-    }
-
-    private fun startPlayer() {
-        mediaPlayer.start()
-        buttonPlay.setImageResource(pause)
-        playerState = STATE_PLAYING
-    }
-
-    private fun pausePlayer() {
-        mediaPlayer.pause()
-        buttonPlay.setImageResource(play)
-        playerState = STATE_PAUSED
-    }
-
-    private fun playbackControl() {
-        when (playerState) {
-            STATE_PLAYING -> {
-                pausePlayer()
-            }
-            STATE_PREPARED, STATE_PAUSED -> {
-                startPlayer()
-            }
-        }
-    }
-
-    override fun onPause() {
-        super.onPause()
-        pausePlayer()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        mediaPlayer.release()
-        mainHandler?.removeCallbacks(runThread)
-    }
-
-    companion object {
-        private const val STATE_DEFAULT = 0
-        private const val STATE_PREPARED = 1
-        private const val STATE_PLAYING = 2
-        private const val STATE_PAUSED = 3
-    }
-
-}
