+++ app/src/main/java/com/sakal/playlistmaker/TrackResponse.kt
@@ -1,4 +1,6 @@
-package com.sakal.playlistmaker.model
+package com.sakal.playlistmaker
+
+import com.sakal.playlistmaker.domain.Track
 
 class TrackResponse(
     val resultCount: Int,
