+++ app/src/main/java/com/sakal/playlistmaker/TracksDiffCallback.kt
@@ -1,7 +1,7 @@
 package com.sakal.playlistmaker
 
 import androidx.recyclerview.widget.DiffUtil
-import com.sakal.playlistmaker.model.Track
+import com.sakal.playlistmaker.domain.Track
 
 class TracksDiffCallback(
     private val trackList: ArrayList<Track>,
