+++ app/src/main/java/com/sakal/playlistmaker/data/MediaPlayerListener.kt
@@ -1,0 +1,9 @@
+package com.sakal.playlistmaker.data
+
+interface MediaPlayerListener {
+    fun setStatePrepared()
+    fun removeHandlersCallbacks()
+    fun setImagePlay()
+    fun setCurrentTimeZero()
+
+}
