+++ app/src/main/java/com/sakal/playlistmaker/domain/PlayerStateListener.kt
@@ -1,0 +1,8 @@
+package com.sakal.playlistmaker.domain
+
+interface PlayerStateListener {
+    fun setStatePrepared()
+    fun removeHandlersCallbacks()
+    fun setImagePlay()
+    fun setCurrentTimeZero()
+}
