+++ app/src/main/java/com/sakal/playlistmaker/presentation/PlayerView.kt
@@ -1,0 +1,6 @@
+package com.sakal.playlistmaker.presentation
+
+interface PlayerView {
+    fun setImage(image: Int)
+    fun setProgressTime(time: String)
+}
